Extract setup wizard startup flow
Move setup wizard startup auto-show and migration auto-scan coordination into a dedicated flow object so the window keeps only UI mechanics while preserving startup timing and existing policies.
