+++ Assets/Tests/Editor/OnionAssemblyDependencyTests.cs
@@ -884,9 +884,19 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
             // Tests that Setup Wizard settings reads run after the synchronous Editor startup hook.
             string setupWizardSource = ReadProductionSource(
                 "Packages/src/Editor/Presentation/Setup/SetupWizardWindow.cs");
+            string setupWizardStartupFlowSource = ReadProductionSource(
+                "Packages/src/Editor/Presentation/Setup/SetupWizardStartupFlow.cs");
 
-            Assert.That(setupWizardSource, Does.Contain("EditorApplication.delayCall += TryShowOnVersionChange;"));
-            Assert.That(setupWizardSource, Does.Not.Contain("\n            TryShowOnVersionChange();"));
+            Assert.That(
+                setupWizardSource,
+                Does.Contain("EditorApplication.delayCall += startupFlow.TryShowOnVersionChange;"));
+            Assert.That(
+                setupWizardStartupFlowSource,
+                Does.Contain("EditorApplication.delayCall += () => _showWindowOnVersionChange();"));
+            Assert.That(
+                setupWizardStartupFlowSource,
+                Does.Contain("EditorApplication.delayCall += () => _showThirdPartyMigrationAutoScan();"));
+            Assert.That(setupWizardSource, Does.Not.Contain("\n            startupFlow.TryShowOnVersionChange();"));
         }
 
         [Test]
