+++ Packages/src/Editor/Presentation/Setup/SetupWizardStartupFlow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62783612b3514214a7a2f1c318d972ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
